+++ produto/back/apiProduto/src/main/java/br/com/tcc/apiproduto/model/Chat.java
@@ -13,7 +13,7 @@ public class Chat {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idChat;
+    private Long idChat;
 
 
     private String mensagens;
